tests/wpt/web-platform-tests/navigation-api/updateCurrentEntry-method/history-pushState.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  navigation.updateCurrentEntry({ state : { data : "value" } });
  assert_equals(navigation.currentEntry.getState().data, "value");
  history.pushState(1, "", "#push");
  assert_equals(navigation.currentEntry.getState(), undefined);
}, "entry.getState() after history.pushState()");
</script>
