LayoutTests/imported/w3c/web-platform-tests/navigation-api/updateCurrentEntry-method/history-pushState-expected.txt


FAIL entry.getState() after history.pushState() assert_equals: expected (undefined) undefined but got (object) object "[object Object]"
